# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Primea Ticketing API",
        "description": "API documentation for Primea Ticketing System - A comprehensive event ticketing platform",
        "contact": {
            "email": "support@primea.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/scans": {
            "post": {
                "tags": [
                    "Scanning"
                ],
                "summary": "Record ticket scan",
                "description": "Records a ticket scan from mobile scanning app. Validates the ticket and updates its status.",
                "operationId": "recordTicketScan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_code",
                                    "scanned_at",
                                    "device_id"
                                ],
                                "properties": {
                                    "qr_code": {
                                        "description": "QR code content from the scanned ticket",
                                        "type": "string",
                                        "example": "TKT123456789"
                                    },
                                    "scanned_at": {
                                        "description": "Timestamp when the ticket was scanned",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-06-15T20:30:00Z"
                                    },
                                    "device_id": {
                                        "description": "Unique identifier of the scanning device",
                                        "type": "string",
                                        "example": "SCANNER_001_GATE_A"
                                    },
                                    "location_hint": {
                                        "description": "Optional location hint where scan occurred",
                                        "type": "string",
                                        "example": "Main Entrance Gate A",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "description": "Optional notes about the scan",
                                        "type": "string",
                                        "example": "Customer had damaged ticket",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valid ticket scan recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Billet validé avec succès"
                                        },
                                        "result": {
                                            "type": "string",
                                            "enum": [
                                                "valid",
                                                "duplicate",
                                                "invalid"
                                            ],
                                            "example": "valid"
                                        },
                                        "scan_id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "ticket": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "TKT123456789"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "issued",
                                                        "used"
                                                    ],
                                                    "example": "used"
                                                },
                                                "event": {
                                                    "properties": {
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Summer Music Festival"
                                                        },
                                                        "venue_name": {
                                                            "type": "string",
                                                            "example": "Central Park"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "ticket_type": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "VIP Access"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "holder": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "john@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "used_at": {
                                                    "type": "string",
                                                    "example": "15/06/2025 20:30:15",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid QR code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR code invalide"
                                        },
                                        "result": {
                                            "type": "string",
                                            "example": "invalid"
                                        },
                                        "scan_id": {
                                            "type": "integer",
                                            "example": 124
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Seuls les organisateurs peuvent enregistrer des scans."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate ticket scan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ce billet a déjà été scanné le 15/06/2025 à 18:45:30"
                                        },
                                        "result": {
                                            "type": "string",
                                            "example": "duplicate"
                                        },
                                        "scan_id": {
                                            "type": "integer",
                                            "example": 125
                                        },
                                        "ticket": {
                                            "description": "Ticket information with duplicate status",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The qr code field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "qr_code": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The qr code field is required."
                                                    }
                                                },
                                                "scanned_at": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The scanned at field is required."
                                                    }
                                                },
                                                "device_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The device id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}